feat: move selection into your ships pane
