+++ packages/client/src/layers/react/components/index.ts
@@ -1,7 +1,6 @@
 import { registerActionQueue } from "./ActionQueue";
 import { registerJoinGame } from "./JoinGame";
 import { registerLoadingState } from "./LoadingState";
-import { registerOptionSelect } from "./OptionSelect";
 import { registerEnemyShip } from "./OverviewComponents/EnemyShip";
 import { registerTopBar } from "./TopBar/TopBar";
 import { registerTurnTimer } from "./TurnTimer";
@@ -11,7 +10,6 @@ export function registerUIComponents() {
   registerLoadingState();
   // registerComponentBrowser();
   registerActionQueue();
-  registerOptionSelect();
   registerTopBar();
   registerEnemyShip();
   registerYourShips();
