+++ packages/client/src/layers/react/styles/global.ts
@@ -66,8 +66,8 @@ export const Button = styled.button<{ isSelected?: boolean; noGoldBorder?: boole
   background: ${({ isSelected }) => `${isSelected ? gold : glass}`};
   border: ${({ noGoldBorder }) => `${noGoldBorder ? "0" : "1"}px solid ${gold}`};
   cursor: pointer;
-  padding: 8px;
-  border-radius: 7px;
+  padding: 6px;
+  border-radius: 6px;
   border-color: ${gold};
   pointer-events: all;
   color: ${darkBrown};
